static function throw and getError
