+++ lib/api/controllers/controller.js
@@ -22,7 +22,8 @@
 const
   { BadRequestError} = require('kuzzle-common-objects').errors,
   _ = require('lodash'),
-  Bluebird = require('bluebird');
+  Bluebird = require('bluebird'),
+  errorsManager = require('../../config/error-codes/throw.js');
 
 // Base class for all API controllers
 class BaseController {
@@ -99,7 +100,8 @@ class BaseController {
    */
 
   throw(...args) {
-    this.kuzzle.throw(this.domain, this.subdomain, ...args);
+
+    errorsManager.throw(this.domain, this.subdomain, ...args);
   }
 
   /**
@@ -110,7 +112,7 @@ class BaseController {
    */
 
   getError(...args) {
-    return this.kuzzle.getError(this.domain, this.subdomain, ...args);
+    return errorsManager.getError(this.domain, this.subdomain, ...args);
   }
 }
