+++ lib/api/kuzzle.js
@@ -49,9 +49,6 @@ const
   Request = require('kuzzle-common-objects').Request,
   InternalEngineBootstrap = require('../services/internalEngine/bootstrap'),
   runShutdown = require('../util/shutdown'),
-  { format } = require('util'),
-  errorCodes = require('../config/error-codes'),
-  { errors } = require('kuzzle-common-objects'),
   Vault = require('./core/vault');
 
 /**
@@ -242,41 +239,6 @@ class Kuzzle extends EventEmitter {
 
         return updated;
       });
-  }
-
-
-  /** Build errors
-  *
-  * @param {...args} ...args
-  * @throws {KuzzleError}
-  */
-
-  getError(...args) {
-
-    const [domain, subdomain, errorName] = args;
-    const placeholders = args.length > 3 ? args.slice(3) : ' ';
-    const name = `${domain}-${subdomain}-${errorName}`;
-    const error = _.get(errorCodes, `${domain}.subdomains.${subdomain}.errors.${errorName}`);
-    if (!error) {
-      return this.getError('internal', 'unexpected', 'unknown_error', ...placeholders);
-    }
-    const code =
-      (errorCodes[domain].code >> 24) |
-      (errorCodes[domain].subdomains[subdomain].code >> 16) |
-      errorCodes[domain].subdomains[subdomain].errors[errorName].code;
-    const message = format(error.message, ...placeholders);
-    return new errors[error.class](message, name, code);
-  }
-
-  /** Manage errors to be thrown
-  *
-  * @param {...args} ...args
-  * @throws {KuzzleError}
-  */
-
-  throw(...args) {
-
-    throw this.getError(...args);
   }
 
   static hash (input) {
